refactor(UI): Remake the flow of the website
- Remove Home page
- Replace the LogsView page with the Query page
- Update the Sidebar
- Update routing
