+++ frontend/src/App.jsx
@@ -1,13 +1,5 @@
 import { BrowserRouter, Routes, Route } from 'react-router-dom'
-import {
-    HomePage,
-    AIPage,
-    LogsViewPage,
-    SettingsPage,
-    AccountPage,
-    NotFoundPage,
-    QueryPage
-} from '@routes'
+import { AIPage, NotFoundPage, LogsViewPage, SettingsPage, AccountPage } from '@routes'
 import { AppLayout } from '@components'
 
 function App() {
@@ -15,12 +7,10 @@ function App() {
         <BrowserRouter>
             <AppLayout>
                 <Routes>
-                    <Route path="/" element={<HomePage />} />
+                    <Route path="/" element={<LogsViewPage />} />
                     <Route path="/ai" element={<AIPage />} />
-                    <Route path="/logs" element={<LogsViewPage />} />
                     <Route path="/settings" element={<SettingsPage />} />
                     <Route path="/account" element={<AccountPage />} />
-                    <Route path="/query" element={<QueryPage />} />
                     <Route path="*" element={<NotFoundPage />} />
                 </Routes>
             </AppLayout>
