--- frontend/src/routes/home.jsx
@@ -1,7 +1,0 @@
-export const HomePage = () => {
-    return (
-        <div className="home-page">
-            <h1 className="text-snow text-2xl font-bold">Home Page</h1>
-        </div>
-    )
-}
