+++ frontend/src/routes/index.js
@@ -1,7 +1,5 @@
 export * from './notFound'
-export * from './home'
-export * from './ai'
-export * from './logsView'
 export * from './account'
 export * from './settings'
-export * from './query'
+export * from './logsView'
+export * from './ai'
